refactor: Room 추천 인터페이스 분리
